rename navigation concepts and move modules to proper locations
