+++ android/features/android-feature-navigation/build.gradle
@@ -32,7 +32,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    api project(path: ':kotlin:concepts:navigation:concept-navigation-client')
+    api project(path: ':kotlin:concepts:concept-navigation')
     api deps.androidx.navigation.fragment
     api deps.androidx.navigation.ui
 }
