+++ android/features/android-feature-navigation/src/main/java/io/matthewnelson/android_feature_navigation/requests/PopBackStack.kt
@@ -17,7 +17,7 @@ package io.matthewnelson.android_feature_navigation.requests
 
 import androidx.annotation.IdRes
 import androidx.navigation.NavController
-import io.matthewnelson.concept_navigation_client.NavigationRequest
+import io.matthewnelson.concept_navigation.NavigationRequest
 
 class PopBackStack(
     @IdRes private val destinationId: Int? = null,
