+++ kotlin/concepts/concept-navigation/src/main/java/io/matthewnelson/concept_navigation/BaseNavigationDriver.kt
@@ -13,11 +13,8 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.concept_navigation_client
+package io.matthewnelson.concept_navigation
 
-/**
- * Extend and create a request specific to the destination desired.
- * */
-abstract class NavigationRequest<T> {
-    abstract fun navigate(controller: T)
+abstract class BaseNavigationDriver<T> {
+    abstract suspend fun submitNavigationRequest(request: NavigationRequest<T>)
 }
