+++ kotlin/features/feature-navigation/build.gradle
@@ -8,6 +8,6 @@ targetCompatibility = JavaVersion.VERSION_1_8
 
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    api project(path: ':kotlin:concepts:navigation:concept-navigation-client')
+    api project(path: ':kotlin:concepts:concept-navigation')
     implementation deps.kotlin.coroutinesCore
 }
